==> notebooks/03A-expert-survey-processing.qmd <==
---
title: Expert Opinion Survey processing
---

```{r}
source("notebooks/initialize-data-analysis.r")
```

# Spaghetti code warning
```{r}
library(tidyverse)
library(ggstats)
library(gtsummary)
```

```{r}
data_survey <- read_csv("data/data-original-processed/survey_results.csv") |>
  select(-comorb_hematological)
```

```{r}
data_survey_factor <- data_survey |>
  mutate(across(everything(), ~ factor(.x, levels = c(1:5))))
```

```{r}
# gglikert(data_survey_factor, sort = "descending", sort_method = "mean")
# gglikert_stacked(data_survey_factor, sort = "descending", sort_method = "mean", add_median_line = TRUE)
```

```{r}
# table_survey_summary <- data_survey |>
#   tbl_summary(type = ~"continuous", statistic = list(all_continuous() ~ "{mean} | {median} ({p25} - {p75})"))

calculate_prop <- function(x) {
  sum(x == 4 | x == 5) / length(x)
}

survey_summary <- data_survey |>
  summarise(across(everything(), list(prop = calculate_prop, median = median, IQR = IQR, mean = mean, sd = sd), .names = "{.fn}.{.col}")) |>
  pivot_longer(
    cols = everything(),
    names_to = c(".value", "variable"),
    names_pattern = "(prop|mean|median|IQR|sd)\\.(.*)"
  )

variables_consensus <- survey_summary |>
  filter(prop >= 0.7) |>
  arrange(desc(prop), desc(mean)) |>
  pull(variable)

data_survey_factor |>
  select(any_of(variables_consensus)) |>
  labelled::set_variable_labels(
    LZD_duration_14 = "LZD duration ≥ 14 days",
    comed_heparin = "Heparin use",
    comed_enoxaparin = "Enoxaparin use",
    infect_sepsis = "Sepsis",
    baseline_CLCR_30 = "Baseline CLCR ≤ 30 mL/min",
    comorb_cirr = "Cirrhosis",
    invasive_CRRT = "CRRT",
    dept_ICU = "ICU admission",
    baseline_PLT_150 = "Baseline PLT ≤ 150 G/L",
    baseline_CLCR = "Baseline CLCR",
    patient_age_65 = "Age ≥ 65 years",
    patient_age = "Age",
  ) |>
  gglikert(sort = "descending", sort_method = "prop")
```

```{r}
data_patient_screen_full <- data_patient_transformed |>
  mutate(
    across(where(is.logical), as.factor)
  )

# Assuming `data_patient_screen_full` is your data frame
single_level_factors <- data_patient_screen_full |>
  purrr::map_lgl(~ is.factor(.x) && length(levels(.x)) == 1)

# Print the names of the variables that are factors with only one level
variables_to_remove <- names(single_level_factors)[single_level_factors]
```

```{r}
variables_to_screen_short <- variables_consensus |>
  str_replace_all("_\\d+", "") |>
  str_replace("LZD_duration", "LZD_duration_14") |>
  str_replace("dept_ICU|dept_ER", "dept_ICU_ER") |>
  unique()
```

```{r}
variables_to_screen_short_binary <- variables_consensus |>
  str_replace_all("_\\d+", "") |>
  str_replace("LZD_duration", "LZD_duration_14") |>
  str_replace("baseline_CLCR", "baseline_CLCR_30") |>
  str_replace("dept_ICU|dept_ER", "dept_ICU_ER") |>
  unique()
```

```{r}
variables_to_screen_binary <- survey_summary |>
  arrange(desc(prop), desc(mean)) |>
  pull(variable) |>
  str_replace_all("_\\d+", "") |>
  str_replace("LZD_duration", "LZD_duration_14") |>
  str_replace("baseline_CLCR", "baseline_CLCR_30") |>
  str_replace("dept_ICU|dept_ER", "dept_ICU_ER") |>
  unique() |>
  head(21)
```

```{r}
variables_to_screen <- survey_summary |>
  arrange(desc(prop), desc(mean)) |>
  pull(variable) |>
  str_replace_all("_\\d+", "") |>
  str_replace("LZD_duration", "LZD_duration_14") |>
  str_replace("dept_ICU|dept_ER", "dept_ICU_ER") |>
  unique() |>
  head(21)

variables_to_force <- variables_to_screen[1:5]

variables_to_force_top_3 <- variables_to_screen[1:3]

variables_out <- survey_summary |>
  filter(!variable %in% variables_to_screen)
```

```{r}
variables_to_force_binary <- survey_summary |>
  arrange(desc(prop), desc(mean)) |>
  pull(variable) |>
  head(5)
```

```{r}
variables_to_screen_all <- survey_summary |>
  arrange(desc(prop), desc(mean)) |>
  filter(!variable == "LZD_route_PO") |>
  pull(variable) |>
  str_replace_all("_\\d+", "") |>
  str_replace("LZD_duration", "LZD_duration_14") |>
  str_replace("dept_ICU|dept_ER", "dept_ICU_ER") |>
  unique() |>
  setdiff(variables_to_remove)
```

```{r}
variables_to_screen_all_binary <- survey_summary |>
  arrange(desc(prop), desc(mean)) |>
  filter(!variable == "LZD_route_PO") |>
  pull(variable) |>
  str_replace_all("_\\d+", "") |>
  str_replace("LZD_duration", "LZD_duration_14") |>
  str_replace("baseline_CLCR", "baseline_CLCR_30") |>
  str_replace("dept_ICU|dept_ER", "dept_ICU_ER") |>
  unique() |>
  setdiff(variables_to_remove)
```

```{r}
save(variables_to_screen, file = "data/results/variables-to-screen.rda")
save(variables_to_screen_short, file = "data/results/variables-to-screen-short.rda")
save(variables_to_screen_short_binary, file = "data/results/variables-to-screen-short-binary.rda")

save(variables_to_screen_binary, file = "data/results/variables-to-screen-binary.rda")
save(variables_to_screen_all, file = "data/results/variables-to-screen-all.rda")
save(variables_to_screen_all_binary, file = "data/results/variables-to-screen-all-binary.rda")

save(variables_to_force, file = "data/results/variables-to-force.rda")
save(variables_to_force_binary, file = "data/results/variables-to-force-binary.rda")
save(variables_to_force_top_3, file = "data/results/variables-to-force-top-3.rda")
```
