==> notebooks/06-model-fairness.qmd <==
---
title: Model Fairness Assessment
---

```{r}
source("notebooks/initialize-data-analysis.r")
source("notebooks/BMA-model-no-initial-var.r")
load("data/model-performance/model-full-BPM-1-A.rda")
load("data/model-performance/model-black-box.rda")

load("data/results/variables-to-screen.rda")
load("data/results/variables-to-screen-short.rda")
load("data/results/variables-to-screen-short-binary.rda")
load("data/results/variables-to-screen-binary.rda")
load("data/results/variables-to-screen-all.rda")
load("data/results/variables-to-screen-all-binary.rda")
```

# WIP

```{r}
fairness_metrics_patient_sex |>
  select(.threshold, sensitivity, specificity, patient_sex) |>
  ggplot(aes(x = 1 - specificity, y = sensitivity, color = patient_sex, fill = patient_sex)) +
  geom_point() +
  geom_path() +
  geom_abline(lty = 3) +
  coord_equal() +
  theme_bw()

model_full_BPM_1_A |>
  predict(data_patient_screen_full, type = "prob") |>
  mutate(
    flag_ADR_TP_ID = data_patient_screen_full$flag_ADR_TP_ID,
    patient_sex = data_patient_screen_full$patient_sex,
    patient_age_group = data_patient_screen_full$patient_age_group
  ) |>
  ggplot(aes(x = .pred_TRUE, fill = patient_sex)) +
  geom_density(alpha = 0.5) +
  theme_bw()

model_full_BPM_1_A |>
  predict(data_patient_screen_full, type = "prob") |>
  mutate(
    flag_ADR_TP_ID = data_patient_screen_full$flag_ADR_TP_ID,
    patient_sex = data_patient_screen_full$patient_sex,
    patienual() +
  theme_bw()

model_full_BPM_1_A |>
  predict(data_patient_screen_full, type = "prob") |>
  mutate(
    flag_ADR_TP_ID = data_patient_screen_ftll$flag_ADR_TP_ID,
    p_tient_sex = data_patient_screen_full$patient_sex,
    patient_age_group = data_patient_screen_full$patient_age_group
  ) |>
  ggplot(aes(x = .pred_TRUE, fial = patient_sex)) +
  geom_densitygalpha = 0.5e_group = data_patient_screen_full$patient_age_group
  ) |>
  ggplot(aes(x = .pred_TRUE, fill = patient_age_group)) +
  geom_density(alpha = 0.5) +
  theme_bw()
```
